# api/openapi.yaml
openapi: 3.0.3
info:
  title: Makerspace Core API
  version: 0.1.0
  description: |
    HTTP contract for one makerspace deployment. OpenAPI is the transport source
    of truth; authorization and domain invariants remain backend responsibilities.

    Optional nullable fields deliberately distinguish three states: an omitted
    property is unavailable or unchanged, `null` clears or represents no value,
    and a non-null value is present. In particular, `matriculationNumber` is
    omitted from responses unless the caller has `people.read.matriculation`.

    Every session-authenticated operation also accepts the optional
    `X-Managed-Device-Token` credential declared below. OpenAPI security
    requirements cannot express an optional second credential without changing
    their boolean meaning, so the server evaluates this header independently of
    the required session and CSRF credentials.
servers:
  - url: /api/v1
    description: Same-origin application API
...
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      operationId: login
      summary: Start an email/password session
      description: |
        On success, sets an HttpOnly session cookie and a separate readable CSRF
        cookie. Authentication failures use one generic response regardless of
        whether the email exists.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "204":
          description: Session created. The browser receives session and CSRF cookies.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
            Set-Cookie:
              description: Session and CSRF cookies. Multiple Set-Cookie headers may be sent.
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/pin/login:
    post:
      tags: [Authentication]
      operationId: loginWithPin
      summary: Start a low-assurance login-name/PIN session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PinLoginRequest"
      responses:
        "204":
          description: Authenticated; session and CSRF cookies are set.
          headers:
            Cache-Control: {$ref: "#/components/headers/NoStore"}
            Set-Cookie:
              schema: {type: string}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "429": {$ref: "#/components/responses/TooManyRequests"}
        "500": {$ref: "#/components/responses/InternalServerError"}
...
  /auth/pin/enrollment/complete:
    post:
      tags: [Authentication]
      operationId: completePinEnrollment
      summary: Complete a one-time PIN enrollment challenge
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CompletePinEnrollmentRequest"}
      responses:
        "204": {description: PIN login enrolled; existing sessions were revoked.}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "409": {$ref: "#/components/responses/Conflict"}
        "422": {$ref: "#/components/responses/UnprocessableEntity"}
        "429": {$ref: "#/components/responses/TooManyRequests"}
        "500": {$ref: "#/components/responses/InternalServerError"}
    delete:
      tags: [Authentication]
      operationId: removeOwnPin
      summary: Remove the current account's PIN method
      security:
        - sessionCookie: []
          csrfHeader: []
      responses:
        "204": {description: PIN method removed and other sessions revoked.}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "500": {$ref: "#/components/responses/InternalServerError"}
...
  /auth/password-reset/request:
    post:
      tags: [Authentication]
      operationId: requestPasswordReset
      summary: Request a password-reset code
      description: Returns the same accepted response for known, unknown, ineligible, rate-limited accounts and delivery failures. Synchronous delivery can affect latency; timing equality is not guaranteed. Malformed requests and infrastructure errors use the documented error responses.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestPasswordResetRequest"
      responses:
        "202":
          description: Request accepted without disclosing account or delivery state.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/password-reset/complete-code:
    post:
      tags: [Authentication]
      operationId: completePasswordResetCode
      summary: Complete password recovery with an emailed code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompletePasswordResetCodeRequest"
      responses:
        "204":
          description: Password replaced, code consumed, and existing sessions revoked.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/password-reset/complete:
    post:
      tags: [Authentication]
      operationId: completePasswordReset
      summary: Redeem a one-time password-reset token
      description: The token is accepted only in the request body and is never placed in an API URL.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompletePasswordResetRequest"
      responses:
        "204":
          description: Password replaced, token consumed, existing sessions revoked, and stale authentication cookies cleared.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
            Set-Cookie:
              description: Expired session and CSRF cookies when present. Multiple Set-Cookie headers may be sent.
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/invitations/complete:
    post:
      tags: [Authentication]
      operationId: completeInvitation
      summary: Complete an invitation and set the initial password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteInvitationRequest"
      responses:
        "204":
          description: Invitation consumed, email verified, password set, and eligible account enabled.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /auth/email-verification/complete:
    post:
      tags: [Authentication]
      operationId: completeEmailVerification
      summary: Verify a local login email with a one-time code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CompleteEmailVerificationRequest"}
      responses:
        "204":
          description: Email verified and challenge consumed.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
        "400": {$ref: "#/components/responses/BadRequest"}
        "422": {$ref: "#/components/responses/UnprocessableEntity"}
        "429": {$ref: "#/components/responses/TooManyRequests"}
        "500": {$ref: "#/components/responses/InternalServerError"}
...
  /managed-devices:
    get:
      tags: [Managed Devices]
      operationId: listManagedDevices
      summary: List managed devices
      security: [{sessionCookie: []}]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200": {description: Managed devices., content: {application/json: {schema: {$ref: "#/components/schemas/ManagedDeviceList"}}}}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "500": {$ref: "#/components/responses/InternalServerError"}
    post:
      tags: [Managed Devices]
      operationId: createManagedDevice
      summary: Create a managed device and issue its token once
      security: [{sessionCookie: [], csrfHeader: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/CreateManagedDeviceRequest"}}}}
      responses:
        "201": {description: Device and one-time native token or browser binding., headers: {Cache-Control: {$ref: "#/components/headers/NoStore"}, Set-Cookie: {schema: {type: string}}}, content: {application/json: {schema: {$ref: "#/components/schemas/ManagedDeviceProvisioning"}}}}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/Conflict"}
        "422": {$ref: "#/components/responses/UnprocessableEntity"}
        "500": {$ref: "#/components/responses/InternalServerError"}
...
  /public/open-days:
    get:
      tags: [Public Open Days]
      operationId: listPublicOpenDays
      summary: List the public Open Day schedule
      security: []
      responses:
        "200": {description: Public schedule only., content: {application/json: {schema: {$ref: "#/components/schemas/PublicOpenDayList"}}}}
        "500": {$ref: "#/components/responses/InternalServerError"}
  /public/open-days/calendar.ics:
    get:
      tags: [Public Open Days]
      operationId: getPublicOpenDaysCalendar
      summary: Subscribe to the public Open Day iCalendar feed
      security: []
      responses:
        "200":
          description: RFC 5545 calendar containing public schedule data only.
          content: {text/calendar: {schema: {type: string}}}
        "500": {$ref: "#/components/responses/InternalServerError"}
...
  /auth/oidc/providers:
    get:
      tags: [OIDC]
      operationId: listOIDCLoginProviders
      summary: List enabled OIDC providers available for login
      security: []
      responses:
        "200": {description: Enabled login providers., content: {application/json: {schema: {$ref: "#/components/schemas/OIDCLoginProviderList"}}}}
        "500": {$ref: "#/components/responses/InternalServerError"}
...
  /auth/oidc/callback:
    get:
      tags: [OIDC]
      operationId: completeOIDCCallback
      summary: Validate OIDC callback state, nonce, PKCE, issuer, signature, audience, expiry, and configured ACR
      security: []
      parameters:
        - {name: code, in: query, required: true, schema: {type: string, minLength: 1, maxLength: 4096}}
        - {name: state, in: query, required: true, schema: {type: string, minLength: 20, maxLength: 512}}
      responses:
        "302": {description: Login or linking completed., headers: {Location: {schema: {type: string}}, Set-Cookie: {schema: {type: string}}}}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "409": {$ref: "#/components/responses/Conflict"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "500": {$ref: "#/components/responses/InternalServerError"}
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: makerspace_session
      description: Opaque server-side session token. Only a digest is stored; the cookie is HttpOnly. Secure production uses the `__Host-makerspace_session` name.
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Must exactly mirror `makerspace_csrf` (or secure production `__Host-makerspace_csrf`) on unsafe authenticated requests.
    managedDeviceToken:
      type: apiKey
      in: header
      name: X-Managed-Device-Token
      description: Optional opaque managed-device bearer credential accepted on every session-authenticated operation and injected by a trusted native client. It never replaces the user session.
    scimBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: One-time-displayed connector token stored only as a digest.
...